made the code more readable and added the backend into the codebase
