+++ package.json
@@ -15,6 +15,7 @@
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "react-router-dom": "^7.6.2",
+    "socket.io-client": "^4.8.1",
     "tailwind": "^4.0.0"
   },
   "devDependencies": {
