+++ src/components/Chat/ChatListItem.jsx
@@ -1,0 +1,11 @@
+
+
+const colors = {
+    lightestBlue: "#e0fbfc",
+    lightBlue: "#c2dfe3",
+    mediumBlue: "#9db4c0",
+    darkBlue: "#5c6b73",
+    darkestBlue: "#253237",
+  };
+
+
