+++ src/services/socket.js
@@ -1,0 +1,4 @@
+import { io } from "socket.io-client";
+import { backendURL } from "./api";
+
+export const socket = io(backendURL);
